refactor: migrate enrich implementation into apps/news_enrich
