+++ apps/news_editorial/src/news_editorial/slugs.py
@@ -1,0 +1,8 @@
+# backend/slugs.py
+import re, unicodedata
+
+def slugify(title: str) -> str:
+    s = unicodedata.normalize("NFKD", title).encode("ascii", "ignore").decode("ascii")
+    s = re.sub(r"[^a-zA-Z0-9]+", "-", s).strip("-").lower()
+    s = re.sub(r"-{2,}", "-", s)
+    return s or "nota"
